Merge dev with master

+++ scripts/uninstall.ksh
@@ -2,6 +2,5 @@ rm -rf ~/.korny
 
 rm -rf .kshrc
 
-if [ -x .kshrc_bak ]; then
-	mv .kshrc_bak .kshrc
-fi
+[[ -f .kshrc_bak ]] && mv .kshrc_bak .kshrc
+
